feat(Samples): move samples to package manager Samples~ directory
The Unity 2019.1 Package Manager allows for samples to be located in
a special `Samples~` root directory which allows the samples to be
optionally included with the package at install time. As the Unity
Input System package requires at least 2019.1 then this functionality
can be used in this package.

The existing `GenericXR` sample has now been moved to this new
directory so it can be optionally added to the project if required.