fix(Transformation): provide deadzone value for float
There was an issue where the minimum float value for an axis
could be so low that it just wasn't reported by Unity as an action
meaning any subsequent action may not get updated due to the value
not being recognised as changed.

This fix adds in an extra deadzone value to the float that can be
used to set the absolute minimum reported value to consider as
zero.